Use `unsigned long` instead of `size_t` and create queue without props.

+++ week_03/heat_stencil/heat_stencil.cl
@@ -1,6 +1,6 @@
-__kernel void calc_temp(__global float const* matrix_a, __global float* matrix_b, size_t n, size_t source_x, size_t source_y) {
-  size_t i = get_global_id(0);
-  size_t j = get_global_id(1);
+__kernel void calc_temp(__global float const* matrix_a, __global float* matrix_b, ulong n, ulong source_x, ulong source_y) {
+  ulong i = get_global_id(0);
+  ulong j = get_global_id(1);
 
   // Center stays constant (the heat is still on).
   if (i == source_x && j == source_y) {
